Refactor main application initialization with modular startup process
- Created new module `app_initialization.py` to handle application startup logic
- Extracted argument parsing, splash screen creation, and initial app state setup
- Simplified main.py by moving complex initialization code to dedicated functions
- Improved type hints and error handling during startup
- Enhanced modularity and readability of application initialization process
